Fix format warning
- size_t is either (depending on compiler / platform) an unsigned long
long, or just an unsigned long
- So simply casting it to an int and end of story ;)

+++ src/Plugin/src/Import/KMLImport/src/KMLImportPlugin.cpp
@@ -205,7 +205,7 @@ bool KMLImportPlugin::import(const QString &filePath, FlightService &flightServi
             }
         } else {
 #ifdef DEBUG
-            qDebug("KMLImportPlugin::import: XML error: %s - number of imported positions: %lu", qPrintable(d->xml.errorString()), aircraft.getPositionConst().count());
+            qDebug("KMLImportPlugin::import: XML error: %s - number of imported positions: %d", qPrintable(d->xml.errorString()), static_cast<int>(aircraft.getPositionConst().count()));
 #endif
             ok = false;
         }
